Move database back out
