+++ neuroscout/models/analysis.py
@@ -1,7 +1,7 @@
 from sqlalchemy.dialects.postgresql import JSONB
 from sqlalchemy.ext.hybrid import hybrid_property
 
-from .database import db
+from ..database import db
 from .utils import copy_row
 
 import datetime
