+++ neuroscout/models/group.py
@@ -1,4 +1,4 @@
-from .database import db
+from ..database import db
 
 
 class GroupPredictor(db.Model):
