+++ neuroscout/populate/ingest.py
@@ -16,8 +16,9 @@ from ..core import cache
 from .utils import hash_stim
 from ..utils.db import get_or_create
 from ..models import (Dataset, Task, Run, Predictor, PredictorEvent,
-                    PredictorRun, Stimulus, RunStimulus,
-                    GroupPredictor, GroupPredictorValue, db)
+                     PredictorRun, Stimulus, RunStimulus,
+                     GroupPredictor, GroupPredictorValue)
+from ..database import db
 from progressbar import progressbar
 from .annotate import PredictorSerializer
