+++ neuroscout/populate/modify.py
@@ -6,7 +6,8 @@ import json
 import re
 from flask import current_app
 from ..models import (Dataset, Task, Run, RunStimulus, Stimulus,
-                      ExtractedFeature, ExtractedEvent, Predictor, db)
+                      ExtractedFeature, ExtractedEvent, Predictor)
+from ..database import db
 from .extract import create_predictors
