+++ neuroscout/resources/analysis/endpoints.py
@@ -1,7 +1,8 @@
 from flask import send_file, current_app
 from flask_apispec import MethodResource, marshal_with, use_kwargs, doc
 from flask_jwt import current_identity
-from ...models import Analysis, Report, db
+from ...models import Analysis, Report
+from ...database import db
 from os.path import exists
 import datetime
 import webargs as wa
