+++ neuroscout/resources/utils.py
@@ -3,7 +3,8 @@ from flask_jwt import jwt_required, current_identity
 from flask_apispec import doc
 import datetime
 from webargs.flaskparser import parser
-from ..models import Analysis, db
+from ..models import Analysis
+from ..database import db
 from marshmallow import ValidationError
 
 import celery.states as states
